+++ shacl-diagram/src/main/java/fr/sparna/rdf/shacl/diagram/PlantUmlPropertyReader.java
@@ -333,10 +333,10 @@ public class PlantUmlPropertyReader {
 	}
 	
 	
-	public Integer readShOrder(Resource constraint) {
+	public Double readShOrder(Resource constraint) {
 		String v = constraintValueReader.readValueconstraint(constraint, SH.order);
 		
-		return (v != null)?Integer.parseInt(v):null;		
+		return (v != null)?Double.parseDouble(v):null;		
 	}
 		
 	public String readShHasValue(Resource constraint) {
